% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redSpruce_ecdfSpatialOffsets}
\alias{redSpruce_ecdfSpatialOffsets}
\title{Empirical cumulative distribution function of spatial offsets}
\format{
A function computed using the ecdf function of the 'stats' package.
\describe{
\item{redSpruce_ecdfSpatialOffsets}{a step function with jumps i/n at observation values, where i is the number of tied spatial offset observations at that value.}
}
}
\usage{
data(redSpruce_ecdfSpatialOffsets)
}
\description{
Empirical cumulative distribution function of contemporary spatial offsets among Southern Appalachian red spruce occurrences
created using the spatOffset function and the 'redSpruce_transClimate_spatialOffsets' data set.
}
\keyword{datasets}
