% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redSpruce_transClimate_spatialOffsets}
\alias{redSpruce_transClimate_spatialOffsets}
\title{Transformed future climate for the Appalachian Blue Ridge Forest ecoregion}
\format{
A data frame with 25 rows and 11 variables:
\describe{
\item{CMD}{Hargreaves climatic moisture deficit}
\item{DD_0}{degree-days below 0°C, chilling degree-days}
\item{DD18}{degree-days above 18°C, cooling degree-days}
\item{eFFP}{the day of the year on which the Frost Free Period ends}
\item{EXT}{extreme maximum temperature over 30 years}
\item{MAR}{mean annual solar radiation}
\item{MSP}{mean summer precipitation}
\item{PAS}{precipitation as snow}
\item{PET}{potential evapotranspiration}
\item{RH}{mean annual relative humidity}
\item{TD}{continentality}
}
}
\source{
\url{https://climatena.ca}
\url{https://envirem.github.io}
}
\usage{
data(redSpruce_transClimate_spatialOffsets)
}
\description{
A data set containing Gradient Forest transformed reference climate normals (1961-1990)
for Southern Appalachian red spruce occurrences (latitudes <40°N).
The original climate variables were transformed using Gradient Forest turnover functions.
The Gradient Forest model was fit to data SNP candidate data for Southern Appalachian red spruce populations.
The transformations provide a common biological scale of compositional turnover (in this case, turnover in allele frequencies),
thereby allowing conversion from multidimensional climatic space to multidimensional genetic space
while selecting and weighting variables such that they best summarize genomic variation.
Please not that due to the transformation the variables are unitless.
}
\keyword{datasets}
