% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redSprucePops_blueRidge_DI_RI}
\alias{redSprucePops_blueRidge_DI_RI}
\title{Donor and Recipient Importance of southern Appalachian red spruce populations}
\format{
A list of two data frames containing donor and recipient importance.
Row names correspond to red spruce population IDs (donor importance) and geographic grid cells of the Appalachian Blue Ridge Forest eco-region.
The corresponding geographic coordinates can be found in XXX.
\describe{
\item{donorImportance}{Donor importance of red spruce populations}
\item{recipientImportance}{Recipient importance of Appalachian Blue Ridge Forest locations}
}
}
\usage{
data(redSprucePops_blueRidge_DI_RI)
}
\description{
Donor and Recipient Importance of southern Appalachian red spruce populations (donors) and Appalachian Blue Ridge Forests (recipients)
calculated using the \code{\link[offsetEnsembleR]{donor.recipientImportance}} function and \code{\link[offsetEnsembleR]{redSprucePops_blueRidge_standardizedOffset}} data.
The 'not-to-exceed' standardized offset threshold was set to 1 sigma.
}
\keyword{datasets}
