% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redSprucePops_blueRidge_rawOffset}
\alias{redSprucePops_blueRidge_rawOffset}
\title{Genomic offset between red spruce populations under current and Appalachian Blue Ridge Forests under future climate}
\format{
A matrix of pairwise raw genomic offsets
\describe{
\item{column names}{Donor population IDs (red spruce populations under current climate)}
\item{row names}{Recipient population IDs (Appalachian Blue Ridge Forest geographic grid cells under future climate)}
}
}
\usage{
data(redSprucePops_blueRidge_rawOffset)
}
\description{
Raw spatio-temporal offsets between southern Appalachian red spruce populations under current
and Appalachian Blue Ridge Forests under future climate.
Calculated using the \code{\link[offsetEnsembleR]{spatTempOffset}} function and \code{\link[offsetEnsembleR]{redSprucePops_transClimate}} and
\code{\link[offsetEnsembleR]{blueRidge_transAlteredClimate}} data sets.
}
\keyword{datasets}
