% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redSprucePops_blueRidge_standardizedOffset}
\alias{redSprucePops_blueRidge_standardizedOffset}
\title{Standardized pairwise genomic offsets between red spruce populations under current and Appalachian Blue Ridge Forests under future climate}
\format{
A matrix of standardized pairwise genomic offsets
\describe{
\item{column names}{Donor population IDs (red spruce populations under current climate)}
\item{row names}{Recipient population IDs (Appalachian Blue Ridge Forest geographic grid cells under future climate)}
}
}
\usage{
data(redSprucePops_blueRidge_standardizedOffset)
}
\description{
Probabilistic re-expression of spatio-temporal offsets between southern Appalachian red spruce populations under current
and Appalachian Blue Ridge Forests under future climate.
Calculated using the \code{\link[offsetEnsembleR]{standardizeOffset}} function and \code{\link[offsetEnsembleR]{redSprucePops_blueRidge_rawOffset}} data.
Standardized offsets probabilistically re-express raw offsets as z-scores based on the empirical distribution of contemporary spatial offsets
\code{\link[offsetEnsembleR]{redSpruce_ecdfSpatialOffsets}}.
}
\keyword{datasets}
