% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redSprucePops_DI_RI}
\alias{redSprucePops_DI_RI}
\title{Donor and Recipient Importance of southern Appalachian red spruce populations}
\format{
A list of two data frames containing donor and recipient importance. Row names correspond to population IDs.
\describe{
\item{donorImportance}{Donor importance}
\item{recipientImportance}{Recipient importance}
}
}
\usage{
data(redSprucePops_DI_RI)
}
\description{
Donor and Recipient Importance of southern Appalachian red spruce populations
calculated using the \code{\link[offsetEnsembleR]{donor.recipientImportance}} function and \code{\link[offsetEnsembleR]{redSprucePops_standardizedOffset}} data.
The 'not-to-exceed' standardized offset threshold was set to 1 sigma.
}
\keyword{datasets}
