% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redSprucePops_rawOffset}
\alias{redSprucePops_rawOffset}
\title{Pairwise genomic offset between red spruce populations under current and future climates}
\format{
A matrix of pairwise raw genomic offsets
\describe{
\item{column names}{Donor population IDs}
\item{row names}{Recipient population IDs}
}

A matrix of pairwise raw genomic offsets
\describe{
\item{column names}{Donor population IDs (red spruce populations under current climate)}
\item{row names}{Recipient population IDs (red spruce populations under future climate)}
}
}
\usage{
data(redSprucePops_rawOffset)

data(redSprucePops_rawOffset)
}
\description{
Raw spatio-temporal offsets between southern Appalachian red spruce populations under current and future climate.
Calculated using the \code{\link[offsetEnsembleR]{spatTempOffset}} function and \code{\link[offsetEnsembleR]{redSprucePops_transClimate}} and
\code{\link[offsetEnsembleR]{redSprucePops_transAlteredClimate}} data sets.

Raw spatio-temporal offsets between southern Appalachian red spruce populations under current and future climate.
Calculated using the \code{\link[offsetEnsembleR]{spatTempOffset}} function and \code{\link[offsetEnsembleR]{redSprucePops_transClimate}} and
\code{\link[offsetEnsembleR]{redSprucePops_transAlteredClimate}} data sets.
}
\keyword{datasets}
