% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatOffset.R
\name{spatOffset}
\alias{spatOffset}
\title{Calculate spatial offsets}
\usage{
spatOffset(
  transEnv,
  nCores = NA,
  nBreaks = NA,
  outpath = NULL,
  returnResult = TRUE
)
}
\arguments{
\item{transEnv}{(dataframe) A dataframe holding transformed environmental variables of all occurrences of a species in a reference geographic region.}

\item{nCores}{(integer) Number of cores to be used in parallel computing. If defined, function will automatically perform in parallel. (default = NA).}

\item{nBreaks}{(integer) Number of chunks to subset rawOffset. Chunks will be sent to single cores. nBreaks should be equal or larger than nCores. If not defined while nCores is defined, nBreaks will be equal to nCores. (default = NA).}

\item{outpath}{(string) Paths to write results. If defined, results will automatically be written to disk. (default = NULL).}

\item{returnResult}{(boolean) if TRUE, data are (also) returned to the R environment independent of whether an outpath is specified. (default = TRUE).}
}
\value{
A function of class ecdf (empirical cumulative distribution function). See help of stats package for more detail.
}
\description{
Calculates contemporary spatial genomic offsets.
}
\examples{
data(redSpruce_transClimate_spatialOffsets)
redSpruce_ecdfSpatialOffsets <- spatOffset(transEnv = redSpruce_transClimate_spatialOffsets)
# Plot eCDF
plot(redSpruce_ecdfSpatialOffsets)
}
