% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatTempOffset.R
\name{spatTempOffset}
\alias{spatTempOffset}
\title{Calculate (spatio-)temporal offsets}
\usage{
spatTempOffset(
  transEnv,
  transAltEnv,
  nCores = NA,
  nBreaks = NA,
  outpath = NULL,
  returnResult = TRUE
)
}
\arguments{
\item{transEnv}{(dataframe) A dataframe holding transformed environmental variables.}

\item{transAltEnv}{(dataframe) A dataframe holding transformed altered environmental variables.}

\item{nCores}{(integer) Number of cores to be used in parallel computing. If defined, function will automatically perform in parallel. (default = NA).}

\item{nBreaks}{(integer) Number of chunks to subset rawOffset. Chunks will be sent to single cores. nBreaks should be equal or larger than nCores. If not defined while nCores is defined, nBreaks will be equal to nCores. (default = NA).}

\item{outpath}{(string) Paths to write results. If defined, results will automatically be written to disk. (default = NULL).}

\item{returnResult}{(boolean) if TRUE, data are (also) returned to the R environment independent of whether an outpath is specified. (default = TRUE).}
}
\value{
A dataframe with pair-wise genomic offsets. Columns in the rawOffset output dataframe correspond to rows of transEnv, rows of output correspond to rows of transAltEnv.
}
\description{
Calculates raw genomic offsets.
}
\examples{
# Example 1: sampled red spruce populations are both donors and recipients
data(redSprucePops_transClimate)
data(redSprucePops_transAlteredClimate)

# Calculate raw spatio-temporal offsets
redSprucePops_rawOffset <- spatTempOffset(transEnv = redSprucePops_transClimate, transAltEnv = redSprucePops_transAlteredClimate, nCores = NA, nBreaks = NA, outpath = NULL)

# View result
redSprucePops_rawOffset[1:10,1:10]

# Get local offsets (only works this way if rows and columns are identical locations)
redSprucePops_localOffset<-diag(as.matrix(redSprucePops_rawOffset))
redSprucePops_localOffset

#' # Example 2: sampled red spruce populations as donors and Blue Ridge ecoregion as recipients
data(redSprucePops_transClimate)
data(blueRidge_transAlteredClimate)

# Calculate raw spatio-temporal offsets
redSprucePops_blueRidge_rawOffset <- spatTempOffset(transEnv = redSprucePops_transClimate, transAltEnv = blueRidge_transAlteredClimate, nCores = NA, nBreaks = NA, outpath = NULL)

# View result
redSprucePops_blueRidge_rawOffset[1:10,1:10]
}
