% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeOffset.R
\name{standardizeOffset}
\alias{standardizeOffset}
\title{Standardize genomic offsets}
\usage{
standardizeOffset(
  rawOffset,
  cdf,
  nCores = NA,
  nBreaks = NA,
  outpath = NULL,
  returnResult = TRUE,
  returnOffsetProb = FALSE
)
}
\arguments{
\item{rawOffset}{(dataframe) A dataframe holding pairwise raw offsets.}

\item{cdf}{(function) Cumulative density function of spatial offsets.}

\item{nCores}{(integer) Number of cores to be used in parallel computing. If defined function will automatically perform in parallel. (default = NA).}

\item{nBreaks}{(integer) Number of chunks to subset rawOffset. Chunks will be sent to single cores. nBreaks should be equal or larger than nCores. If not defined while nCores is defined, nBreaks will be equal to nCores. (default = NA).}

\item{outpath}{(string) Paths to write results. If defined, results will automatically be written to disk. (default = NULL).}

\item{returnResult}{(boolean) if TRUE, data are (also) returned to the R environment independent of whether an outpath is specified. (default = TRUE).}

\item{returnOffsetProb}{(boolean) if TRUE, a matrix with offset probabilities (drawn from cfd) is also returned.
Caution:will double the required disk space and memory! (default = FALSE).}
}
\value{
A dataframe with offset z-scores (sigma). Dataframe columns and rows are identical with those of the rawOffset input file.
Optionally, the offset probabilities calculated in step 1 can be returned as well.
}
\description{
Standardizes raw genomic offsets based on methods presented by Lachmuth et al. (in review). First, offsets are
expressed  as probabilities of the distribution of spatial genomic offsets obtained using the 'spatOffset' function. Second, the
probabilities are converted to percentiles of a chi distribution with one degree of freedom (quantile normalization).
}
\examples{
# Sampled red spruce populations as donors and Blue Ridge ecoregion as recipients
data(redSprucePops_blueRidge_rawOffset)
data(redSpruce_ecdfSpatialOffsets)

# Standardize raw spatio-temoral offsets
redSprucePops_blueRidge_standardizedOffset <-
standardizeOffset(rawOffset = redSprucePops_blueRidge_rawOffset,
cdf = redSpruce_ecdfSpatialOffsets)

# View results
redSprucePops_blueRidge_standardizedOffset[1:10,1:10]
}
